% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_file.R
\name{add_anchor_file}
\alias{add_anchor_file}
\title{Provide the table of cross references}
\usage{
add_anchor_file(fname = "XREFS.rda")
}
\arguments{
\item{fname}{Name of rda file with anchor table (def: XREFS.rda)}
}
\description{
The {xref} system scans HTML files to locate cross-reference
anchors and stores these anchors in a data frame. `xf()` looks
to this anchor data frame to resolve cross references, turning
the ID into a link and label.
}
