% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors_make.R
\name{anchor_definition}
\alias{anchor_definition}
\title{Create cross-reference anchors}
\usage{
anchor_definition(word, for_show = word)
}
\description{
The {xref} system makes use of cross-reference anchors
embedded in document files. In Quarto, these are created with
labels like `#fig-my-id`. These functions allow you to create them
from R.
}
