% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_from_qmd}
\alias{extract_from_qmd}
\alias{find_terms}
\title{Extract instances of xf() calls from a set of qmd files.}
\usage{
extract_from_qmd(file_vector)

find_terms(file_vector, terms)
}
\arguments{
\item{file_vector}{String vector containing file names}

\item{A}{character string vector with the words to look for in each file}
}
\value{
A data frame where each row is one instance.
}
\description{
Extract instances of xf() calls from a set of qmd files.
}
