% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xf_HTML_to_XREFS.R
\name{make_xf_index}
\alias{make_xf_index}
\alias{xref_html_files}
\alias{xref_make_index}
\alias{xref_fingers_find}
\alias{xref_sections_find}
\alias{xref_figures_find}
\alias{xref_exercises_find}
\alias{xref_tables_find}
\alias{xref_callouts_find}
\title{Make the XREFS.rda file}
\usage{
xref_html_files(save = FALSE)

xref_make_index(htmlname)

xref_fingers_find(Lines)

xref_sections_find(Lines)

xref_figures_find(Lines)

xref_exercises_find(Lines)

xref_tables_find(Lines)

xref_callouts_find(Lines)
}
\arguments{
\item{save}{Flag to save the index to a file. Only set to `FALSE` for debugging.}
}
\description{
In the xf cross-referencing system, which is an alternative
to the pandoc cross referencing, the author inserts markup like
`xf_definition("new word")` or `xf("fig-my-fig")`
Such commands insert content into the HTML files created by quarto.
These functions scan the HTML files and use the xf_ content to create an
index, by default called `XREFS.rda`, which provides the basis for `xf("fig-my-fig")` and such
The user-interface to this HTML to index translation is
}
