% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xf.R
\name{xf}
\alias{xf}
\alias{xf_anchor}
\alias{xf_definition}
\alias{xf_to_def}
\alias{xf_new_words}
\title{Refer to a cross-reference}
\usage{
xf(id, suppress_label = FALSE, label = NULL)

xf(id, suppress_label = FALSE, label = NULL)

xf_anchor(ID, label = "unlabelled")

xf_definition(word, text = word)

xf_to_def(word, label = word)

xf_new_words(chapter)
}
\arguments{
\item{id}{Character string naming the ID to make a cross reference to, e.g. "fig-my-id".}

\item{label}{Character string. Disregard the cross-reference title and use this label instead.}

\item{just_number}{Logical (default FALSE) indicating whether the use
the full title of the cross-reference or just the "number."}
}
\description{
In Quarto, you create cross-references with markup like
`label: fig-my-id`, and reference the location using the
notation `@fig-my-id`. Unfortunately, this makes it hard
to do cross-referencing between different projects or files
within a project (unless "book" mode for Quarto is being used).
}
\details{
With `xf()`, you reference using R syntax: `r xf("fig-my-id")`.
}
